+++ storage/innobase/include/clone0monitor.h
@@ -53,7 +53,8 @@ class Clone_Monitor {
         m_estimate_bytes_left(),
         m_work_bytes_left(),
         m_cur_phase(NOT_STARTED) {
-    m_pfs_data_chunk_size = ut_2_exp(PFS_DATA_CHUNK_SIZE_POW2);
+    m_pfs_data_chunk_size =
+        static_cast<uint>(ut_2_exp(PFS_DATA_CHUNK_SIZE_POW2));
   }
 
   /** Destructor. */
